# docs/swagger.yaml
basePath: /api
definitions:
  common.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  dto.AuthRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.AuthResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      token:
        type: string
    type: object
  dto.CreateTaskRequest:
    properties:
      task:
        example: Buy milk
        maxLength: 20
        type: string
    required:
    - task
    type: object
  dto.CreateUserRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  dto.CreateUserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
    type: object
  dto.DeleteTaskResponse:
    properties:
      message:
        example: Task deleted successfully
        type: string
    type: object
  dto.DeleteUserResponse:
    properties:
      message:
        type: string
    type: object
  dto.GetTaskResponse:
    properties:
      id:
        example: 1
        type: integer
      status:
        example: pending
        type: string
      task:
        example: Buy milk
        type: string
    type: object
  dto.ListTasksResponse:
    properties:
      tasks:
        items:
          $ref: '#/definitions/dto.GetTaskResponse'
        type: array
    type: object
  dto.UpdatePasswordRequest:
    properties:
      id:
        type: integer
      new_password:
        minLength: 6
        type: string
      old_password:
        type: string
    required:
    - id
    - new_password
    - old_password
    type: object
  dto.UpdatePasswordResponse:
    properties:
      message:
        type: string
    type: object
  dto.UpdateStatusRequest:
    properties:
      status:
        enum:
        - pending
        - in-progress
        - completed
        example: completed
        type: string
    required:
    - status
    type: object
  dto.UpdateStatusResponse:
    properties:
      message:
        example: status updated
        type: string
    type: object
  task.Task:
    properties:
      created_at:
        example: "2025-08-27 10:35:16.263"
        type: string
      id:
        example: 1
        minimum: 1
        type: integer
      status:
        example: pending
        type: string
      task:
        example: Buy milk
        type: string
    required:
    - status
    - task
    type: object
host: localhost:8080
info:
  contact: {}
  description: API server for managing tasks and users in the TaskFlow application.
  title: TaskFlow API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: User login
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "409":
          description: Email already exists
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /tasks:
    get:
      consumes:
      - application/json
      description: Get a list of all tasks
      produces:
      - application/json
      responses:
        "200":
          description: List of tasks retrieved successfully
          schema:
            $ref: '#/definitions/dto.ListTasksResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: List all tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Create a new task with title and description
      parameters:
      - description: Task to create
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/task.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Create a new task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      description: Delete a task by its ID
      parameters:
      - description: Task ID
        example: 1
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Task deleted successfully
          schema:
            $ref: '#/definitions/dto.DeleteTaskResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Delete a task
      tags:
      - tasks
    get:
      description: Returns details of a specific task by ID
      parameters:
      - description: Task ID (>=1)
        example: 1
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Task retrieved successfully
          schema:
            $ref: '#/definitions/dto.GetTaskResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get a task by ID
      tags:
      - tasks
  /tasks/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update the status field of a task by its ID
      parameters:
      - description: Task ID
        example: 1
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: Status payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Status updated successfully
          schema:
            $ref: '#/definitions/dto.UpdateStatusResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Update task status
      tags:
      - tasks
  /users/account:
    delete:
      consumes:
      - application/json
      description: Delete user account (requires authentication)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DeleteUserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user account
      tags:
      - users
  /users/password:
    patch:
      consumes:
      - application/json
      description: Update user's password (requires authentication)
      parameters:
      - description: Password update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdatePasswordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user password
      tags:
      - users
swagger: "2.0"